структура изменена
