+++ LR/07/src/tcpserver.cpp
@@ -1,6 +1,5 @@
 #include "tcpserver.h"
-#include <QDebug>
-#include <QCoreApplication>
+
 /**
  * @brief Конструктор класса, в котором создается объект класса QTcpServer
  * Сервер включается и ждет новых соединений.
@@ -12,12 +11,13 @@ TcpServer::TcpServer(QObject *parent)
 
     connect(mTcpServer, &QTcpServer::newConnection, this, &TcpServer::slotNewConnection);
 
-    if (!mTcpServer->listen(QHostAddress::Any, 10000)) {
+    if (!mTcpServer->listen(QHostAddress::LocalHost, 10000)) {
         qDebug() << "Server is not started";
     } else {
         qDebug() << "Server is started";
     }
 }
+
 /**
  * @brief Метод отвечающий за подключение клиента к серверу
  * @return void
@@ -27,24 +27,25 @@ void TcpServer::slotNewConnection()
     mTcpSocket = mTcpServer->nextPendingConnection();
 
     mTcpSocket->write("New connection!");
-    connect(mTcpSocket, &QTcpSocket::readyRead, this, &TcpServer::slotReadingDataJson);
-        SendToClient(false);
 
+    connect(mTcpSocket, &QTcpSocket::readyRead, this, &TcpServer::slotReadingDataJson);
     connect(mTcpSocket, &QTcpSocket::disconnected, this, &TcpServer::slotClientDisconnected);
 }
-/**
- * @brief Метод, считывающий количество байтов, отличных от нуля, передаваемых серверу
- * @return void
- */
-void TcpServer::slotServerRead()
-{
-    while (mTcpSocket->bytesAvailable() > 0) {
 
-        QByteArray array = mTcpSocket->readAll();
+///**
+// * @brief Метод, считывающий количество байтов, отличных от нуля, передаваемых серверу
+// * @return void
+// */
+//void TcpServer::slotServerRead()
+//{
+//    while (mTcpSocket->bytesAvailable() > 0) {
 
-        mTcpSocket->write(array);
-    }
-}
+//        QByteArray array = mTcpSocket->readAll();
+
+//        mTcpSocket->write(array);
+//    }
+//}
+
 /**
  * @brief метод выключает сервер.
  * @return void
@@ -53,15 +54,23 @@ void TcpServer::slotClientDisconnected()
 {
     mTcpSocket->close();
 }
+
 /**
  * @brief Метод отправляет 1 или 0 клиенту
  * @param int answer - ответ, отправляемый клиенту(1 или 0)
  * @return void
  */
-void TcpServer::slotSendToClient(int answer)
+void TcpServer::sendToClient(bool answer)
 {
-    mTcpSocket->write((char*) &answer, sizeof(int));
+   if (answer) {
+       mTcpSocket->readAll();
+       mTcpSocket->write("1");
+   } else {
+       mTcpSocket->readAll();
+       mTcpSocket->write("0");
+   }
 }
+
 /**
  * @brief Метод получает данные от клиента в формате json
  * @return void
@@ -74,32 +83,18 @@ void TcpServer::slotReadingDataJson()
         mTcpSocket->waitForConnected(500);
         data = mTcpSocket->readAll();
         docJson = QJsonDocument::fromJson(data, &docJsonError);
-        if (docJsonError.errorString().toInt()==QJsonParseError::NoError) {
-            if (docJson.object().value("code").toString() == "content") {
-                qDebug() << "ReadingDataJson() - work";
-            } else {
-                qDebug() << "ReadingDataJson()- don't work";
+
+        // TODO вместо docJson.object() нужно вызывать функцию конфертации json в map-объект (задача #47)
+        if (docJsonError.errorString().toInt() == QJsonParseError::NoError) {
+            try {
+                lab = new StrategyLab(docJson.object());
+//                this->sendToClient(lab->check());
+                delete lab;
+            } catch (QString err) {
+                // TODO добавление текста ошибки к ответу клиенту (задача #195)
+                qDebug() << err;
             }
         }
     }
 }
-/**
- * @brief Метод отправляет клиенту 0 либо 1
- * @return void
- */
-void TcpServer::SendToClient(bool answer)
-{
-
-     if (answer) {
-          mTcpSocket->readAll();
-          mTcpSocket->write("1");
-     } else
-     {
-         mTcpSocket->readAll();
-         mTcpSocket->write("0");
-     }
-
-}
-
-
