--- TRPO/main.cpp
@@ -1,11 +1,0 @@
-#include <QCoreApplication>
-#include "tcpserver.h"
-
-int main(int argc, char *argv[])
-{
-    QCoreApplication a(argc, argv);
-
-    TcpServer * server = new TcpServer();
-
-    return a.exec();
-}
